Ajout d'un produit au menu du restau #34

+++ src/main/java/sophiatech/Product.java
@@ -5,9 +5,9 @@ import java.util.Objects;
 public class Product {
     private Restaurant restaurant;
     private String name;
-    private int price;
+    private double price;
 
-    public Product(Restaurant restaurant, String name, int price) {
+    public Product(Restaurant restaurant, String name, double price) {
         this.restaurant = restaurant;
         this.name = name;
         this.price = price;
@@ -19,7 +19,7 @@ public class Product {
         this(null, name, price);
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return this.price;
     }
